+++ internal/logger/test.go
@@ -6,7 +6,14 @@ import (
 )
 
 func NewTest() Logger {
-	zapConfig := zap.NewDevelopmentConfig()
+	zapConfig := zap.Config{
+		Level:            zap.NewAtomicLevelAt(zap.DebugLevel),
+		Development:      false,
+		Encoding:         "console",
+		EncoderConfig:    zap.NewDevelopmentEncoderConfig(),
+		OutputPaths:      []string{"stderr"},
+		ErrorOutputPaths: []string{"stderr"},
+	}
 	zapConfig.EncoderConfig.EncodeLevel = zapcore.CapitalColorLevelEncoder
 	logger, _ := zapConfig.Build()
 	zap.RedirectStdLog(logger)