+++ CHANGELOG.md
@@ -4,9 +4,11 @@
 
 - Breaking: Resolver caches are now stored in PostgreSQL. See [docs/build.md](./docs/build.md) for prerequisite instructions. (#271)
 - Fix: SevenTV emotes now resolve correctly. (#281)
+- Fix: YouTube videos are no longer resolved as channels. (#284)
 - Dev: Improve BetterTTV emote tests. (#282)
 - Minor: BetterTTV cache key changed from plural to singular form. (#282)
 - Dev: Improve Twitch.tv clip tests. (#283)
+- Dev: Improve YouTube tests. (#284)
 
 ## 1.2.3